pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synfin"
version = "0.1.0"
description = "Synthetic Financial Data Generation using TimeGAN, Diffusion Models, and VAE+Copula"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "synfin contributors"}
]
keywords = ["synthetic data", "finance", "time series", "generative models", "deep learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=12.0",
    "yfinance>=0.2",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "mplfinance>=0.12.9b0",
    "seaborn>=0.12",
    "plotly>=5.15",
    "pyyaml>=6.0",
    "tensorboard>=2.13",
    "tqdm>=4.65",
    "ta>=0.11",
    "statsmodels>=0.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.1",
    "mypy>=1.5",
    "jupyter>=1.0",
    "ipykernel>=6.25",
]
wandb = ["wandb>=0.15"]

[project.scripts]
synfin-download = "synfin.data.download:main"
synfin-train = "scripts.train:main"
synfin-generate = "scripts.generate:main"
synfin-evaluate = "scripts.evaluate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
